Mixed use support
You can now optionally configure a "no store" domain.  When inbucket
receives a message destined for a user at that domain, it will accept
the message but not store it to disk.  This allows the same instance
of Inbucket to be shared by people who wish to view email content and
those who want to load test.